writecard_tool/requirements.txt
nfcpy==1.0.3
ndeflib==1.0.3


